Rename and reorganise checkout wizard components
- Hopefully this makes things a little bit more straightforward to understand.
- The checkout wizard is a 6-step process and uses a Wizard component from the react-use-wizard package
- The steps directory contains each step in the checkout process (e.g. date selection, product selection, etc.)
- the wizard-parts directory contains parts of the checkout wizard component, like the WizardHeader which shows the step title (e.g. Select your products), the WizardContainer which houses the Wizard component and all the steps involved
- You can technically pass WizardButtons as a component to Wizard, but since you'll need to disable/enable the Continue button depending on whether a customer has fulfilled a certain task (like signing a waiver, for example), I think it would  be best to just add the buttons to each step individually. There may be some merit to making a reusable 'Continue' button component, but for now, this is what we've got
